Implemented CRUD Operation for "Taluka" and "Town" Entity

+++ src/main/java/com/example/repository/TalukaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.entity.Taluka;
 
 @Repository
-public interface TalukaRepository extends JpaRepository<Taluka, Integer>{
+public interface TalukaRepository extends JpaRepository<Taluka, Long>{
 
 }
