+++ src/main/java/com/example/repository/TownRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.entity.Town;
 
 @Repository
-public interface TownRepository extends JpaRepository<Town, Integer>{
+public interface TownRepository extends JpaRepository<Town, Long>{
 
 }
