+++ src/main/java/com/example/service/TalukaFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Taluka;
 
 public interface TalukaFetchService {
-	Taluka fetchById(int id);
+	Taluka fetchById(Long id);
 
 }
