+++ src/main/java/com/example/service/TalukaUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Taluka;
 
 public interface TalukaUpdateService {
-	Taluka updateData(Taluka taluka , int id);
+	Taluka updateData(Taluka taluka , Long id);
 
 }
