+++ src/main/java/com/example/service/TownUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Town;
 
 public interface TownUpdateService {
-	Town updateData(Town town , int id);
+	Town updateData(Town town , Long id);
 
 }
